Move extensions to common file
